Make autocomplete an ES module
I was hesitant to modify the file at first. However, since the
levenshtein lib had to be modified, I might as well just modify this
too. In practice, this lib will likely never get updated, but if it
does, then it's just a matter of changing 1 line to make it work again.